refactor: move test files
place test files next to their sut
