+++ src/lib/services/fileChange/fileChangeHandlers/jsonFileChangeHandler.test.ts
@@ -2,14 +2,14 @@ import * as assert from 'assert';
 import sinon from 'sinon';
 import { Uri } from 'vscode';
 
-import { ChainType } from '../../lib/enums/chainType';
-import ModuleChainManager from '../../lib/modules/moduleChainManager';
-import ReadJsonFileModule from '../../lib/modules/readJsonFile/readJsonFileModule';
-import TranslationModule from '../../lib/modules/translation/translationModule';
-import JsonFileChangeHandler from '../../lib/services/fileChange/fileChangeHandlers/jsonFileChangeHandler';
-import FileWatcherCreator from '../../lib/services/fileChange/fileWatcherCreator';
-import FileLockStoreStore from '../../lib/stores/fileLock/fileLockStore';
-import filePathUtilities from '../../lib/utilities/filePathUtilities';
+import { ChainType } from '../../../enums/chainType';
+import ModuleChainManager from '../../../modules/moduleChainManager';
+import ReadJsonFileModule from '../../../modules/readJsonFile/readJsonFileModule';
+import TranslationModule from '../../../modules/translation/translationModule';
+import FileLockStoreStore from '../../../stores/fileLock/fileLockStore';
+import filePathUtilities from '../../../utilities/filePathUtilities';
+import FileWatcherCreator from '../fileWatcherCreator';
+import JsonFileChangeHandler from './jsonFileChangeHandler';
 
 suite('JsonFileChangeHandler', () => {
   test('should initialize moduleChainManager and register chain', () => {
