+++ src/lib/services/fileChange/fileChangeHandlers/typeScriptFileChangeHandler.ts
@@ -1,13 +1,13 @@
 import * as Sentry from '@sentry/node';
 import { Uri } from 'vscode';
 
-import { ChainType } from '../enums/chainType';
-import ActionModule from '../interfaces/actionModule';
-import FileChangeHandler from '../interfaces/fileChangeHandler';
-import ModuleContext from '../interfaces/moduleContext';
-import I18nextScannerModule from '../modules/i18nextScanner/i18nextScannerModule';
-import ModuleChainManager from '../modules/moduleChainManager';
-import FileContentStore from '../stores/fileContent/fileContentStore';
+import { ChainType } from '../../../enums/chainType';
+import ActionModule from '../../../interfaces/actionModule';
+import FileChangeHandler from '../../../interfaces/fileChangeHandler';
+import ModuleContext from '../../../interfaces/moduleContext';
+import I18nextScannerModule from '../../../modules/i18nextScanner/i18nextScannerModule';
+import ModuleChainManager from '../../../modules/moduleChainManager';
+import FileContentStore from '../../../stores/fileContent/fileContentStore';
 
 export default class TypeScriptFileChangeHandler implements FileChangeHandler {
   private static i18nextScannerModule: I18nextScannerModule;
