+++ src/lib/services/fileIo/fileReader.test.ts
@@ -2,7 +2,7 @@ import * as assert from 'assert';
 import * as mock from 'mock-fs';
 import { Uri } from 'vscode';
 
-import FileReader from '../../lib/services/fileIo/fileReader';
+import FileReader from './fileReader';
 
 suite('FileReader Tests', () => {
   setup(() => {
