+++ src/lib/services/translate/deeplService.test.ts
@@ -4,10 +4,10 @@ import assert from 'assert';
 import * as deepl from 'deepl-node';
 import sinon from 'sinon';
 
-import TranslationModuleConfiguration from '../../lib/entities/configuration/modules/translationModule/translationModuleConfiguration';
-import DeeplService from '../../lib/services/translate/deeplService';
-import ConfigurationStore from '../../lib/stores/configuration/configurationStore';
-import ConfigurationStoreManager from '../../lib/stores/configuration/configurationStoreManager';
+import TranslationModuleConfiguration from '../../entities/configuration/modules/translationModule/translationModuleConfiguration';
+import ConfigurationStore from '../../stores/configuration/configurationStore';
+import ConfigurationStoreManager from '../../stores/configuration/configurationStoreManager';
+import DeeplService from './deeplService';
 
 suite('DeeplService', () => {
   let deeplService: DeeplService;
