+++ src/lib/stores/configuration/configurationStoreManager.test.ts
@@ -2,10 +2,10 @@ import * as assert from 'assert';
 import sinon from 'sinon';
 import * as vscode from 'vscode';
 
-import ExtensionConfiguration from '../../lib/entities/configuration/extensionConfiguration';
-import TranslationModuleConfiguration from '../../lib/entities/configuration/modules/translationModule/translationModuleConfiguration';
-import ConfigurationStore from '../../lib/stores/configuration/configurationStore';
-import ConfigurationStoreManager from '../../lib/stores/configuration/configurationStoreManager';
+import ExtensionConfiguration from '../../entities/configuration/extensionConfiguration';
+import TranslationModuleConfiguration from '../../entities/configuration/modules/translationModule/translationModuleConfiguration';
+import ConfigurationStore from './configurationStore';
+import ConfigurationStoreManager from './configurationStoreManager';
 
 suite('ConfigurationStoreManager', () => {
   let getConfigurationStub: sinon.SinonStub;
